(#7962) Extracted logic for parsing values in seconds into a util module.
Pulled the logic for specifying time in units of hours, days, and
years out of the certificate factory and into `Puppet::Util::Seconds`.
This allows it to be easily reused for other settings and tested
independently. Updated the `configtimeout` setting module to leverage
the module's error handling.

Note: the format and available units were not changed.