(#7962) Added check for certificate expiration.
Added the `certificate_expire_warning` config option to allow adjusting
the notification lead time of an expiring certificate. Can be specified
in the same format as the `ca_ttl` option.

Currently, the check only occurs when the cert is first initialized,
and the only action that is taken is logging a warning with the type of
cert and its expiration date.