(#7962) Added check for certificate expiration.
Certificate now has a `check_expiration()` function that will log a
warning if its expiration date is approaching with its name and the
date.

Added the `certificate_expire_warning` config option to allow adjusting
the notification lead time of an expiring certificate. Can be specified
in the same format as the `ca_ttl` option.