(#7962) Added network response proxy class.
The response object in `Net::HTTP` doesn't provide access to much about
the reuquest/response, which is a problem when the connection is
abstracted and only the response is handed back. The instances of this
class act as `Net::HTTPResponse*` objects but also provide access to the
session and most importantly the peer/ca certificates used in the
request. The peer cert is made avaiable through the `peer_cert` attr,
mimicking `HTTPClient`'s response interface.